// src/app/models/ewm-aturized-status.model.ts
export interface EWM_Aturized_Status {
  idewm_Aturized_Status?: number;
  codeAtorisedStatus: string;
  valueAtorisedStatus: string;
  okStatus?: boolean;
  nokStatus?: boolean;
  allStatus?: boolean;
}

export interface EWM_Authorized_Status {
  idewm_Authorized_Status?: number;
  codeAuthorizedStatus: string;
  valueAuthorizedStatus: string;
  okStatus: boolean;
  nokStatus: boolean;
  allStatus: boolean;
}